isobic/requirements.txt
pandas==2.2.0
PyPDF2==3.0.1
Requests==2.31.0
tabula_py==2.6.0
